requirements.txt
aiosqlite==0.21.0
annotated-types==0.7.0
asyncpg==0.30.0
black==25.9.0
boto3==1.40.58
botocore==1.40.58
click==8.3.0
iniconfig==2.3.0
iso8601==2.1.0
jmespath==1.0.1
mccabe==0.7.0
mypy_extensions==1.1.0
packaging==25.0
pathspec==0.12.1
platformdirs==4.5.0
pluggy==1.6.0
pycodestyle==2.14.0
pydantic==2.12.3
pydantic_core==2.41.4
pyflakes==3.4.0
Pygments==2.19.2
pypika-tortoise==0.6.2
pytest==8.4.2
pytest-asyncio==0.22.0
python-dateutil==2.9.0.post0
python-dotenv==1.1.1
pytokens==0.2.0
pytz==2025.2
redis==7.0.0
s3transfer==0.14.0
six==1.17.0
structlog==25.4.0
tortoise-orm==0.25.1
typing-inspection==0.4.2
typing_extensions==4.15.0
urllib3==2.5.0
